adding tests for outpartition index computer

+++ src/cpp/main/algos/clever_pagerank/clever_pagerank_node.h
@@ -27,7 +27,7 @@ public:
   void final();
 
   void setNumberNeighbors(vector<int>*);
-  void setOutPartitions(vector<set<long> >*);
+  void setOutPartitions(vector<set<int> >*);
   void setInverseNodeBounds(unordered_map<long, std::pair<long, long> >*);
   void setInverseOutEdges(vector<long>*);
   void setOutputFileName(string);
@@ -37,7 +37,7 @@ public:
   void readInverseNodeBounds(string);
   void readInverseOutEdges(string);
 private:
-  vector<set<long> >* outPartitions;
+  vector<set<int> >* outPartitions;
   unordered_map<long, std::pair<long, long> >* inverseNodeBounds;
   vector<long>* inverseOutEdges;
   vector<int>* numNeighbors;
